+++ tests/Alert/SessionAlertStreamTest.php
@@ -10,6 +10,7 @@
 
 namespace UserFrosting\Tests\Alert;
 
+use Illuminate\Session\ArraySessionHandler;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
@@ -23,7 +24,6 @@ class SessionAlertStreamTest extends TestCase
     use MockeryPHPUnitIntegration;
 
     protected string $key = 'alerts';
-
     protected string $session_id = 'foo123';
 
     public function testConstructor(): void
@@ -39,23 +39,6 @@ class SessionAlertStreamTest extends TestCase
     /**
      * @depends testConstructor
      */
-    public function testSetTranslator(): void
-    {
-        $translator = m::mock(Translator::class);
-        $session = m::mock(Session::class);
-        $stream = new SessionAlertStream($this->key, $translator, $session);
-
-        $this->assertSame($translator, $stream->translator());
-
-        $translator2 = m::mock(Translator::class);
-        $this->assertNotSame($translator, $translator2);
-        $this->assertInstanceOf(SessionAlertStream::class, $stream->setTranslator($translator2)); // @phpstan-ignore-line
-        $this->assertSame($translator2, $stream->translator());
-    }
-
-    /**
-     * @depends testConstructor
-     */
     public function testAddMessage(): void
     {
         // Build Mock
@@ -64,8 +47,9 @@ class SessionAlertStreamTest extends TestCase
 
         // Set expectations
         $message = [
-            'type'    => 'success',
-            'message' => 'foo',
+            'type'         => 'success',
+            'message'      => 'foo',
+            'placeholders' => [],
         ];
         $session->shouldReceive('get')->with($this->key)->once()->andReturn(false);
         $session->shouldReceive('set')->with($this->key, [$message])->once()->andReturn(null);
@@ -86,8 +70,9 @@ class SessionAlertStreamTest extends TestCase
 
         // Set expectations
         $message = [
-            'type'    => 'success',
-            'message' => 'foo',
+            'type'         => 'success',
+            'message'      => 'foo',
+            'placeholders' => [],
         ];
         $session->shouldReceive('get')->with($this->key)->once()->andReturn([$message]);
         $session->shouldReceive('set')->with($this->key, [$message, $message])->once()->andReturn(null);
@@ -117,25 +102,13 @@ class SessionAlertStreamTest extends TestCase
     /**
      * @depends testConstructor
      */
-    public function testAddMessageTranslatedWithNoTranslator(): void
-    {
-        $session = m::mock(Session::class);
-        $stream = new SessionAlertStream($this->key, null, $session);
-
-        $this->expectException(\RuntimeException::class);
-        $stream->addMessageTranslated('success', 'foo', []);
-    }
-
-    /**
-     * @depends testConstructor
-     */
     public function testAddMessageTranslated(): void
     {
         // Build Mock
         $translator = m::mock(Translator::class);
         $session = m::mock(Session::class);
 
-        //
+        // Data
         $key = 'FOO';
         $placeholder = ['key' => 'value'];
         $result = 'Bar';
@@ -143,15 +116,20 @@ class SessionAlertStreamTest extends TestCase
         // Set expectations
         $translator->shouldReceive('translate')->with($key, $placeholder)->andReturn($result);
         $message = [
-            'type'    => 'success',
-            'message' => $result,
+            'type'         => 'success',
+            'message'      => 'FOO',
+            'placeholders' => ['key' => 'value'],
         ];
         $session->shouldReceive('get')->with($this->key)->once()->andReturn(false);
         $session->shouldReceive('set')->with($this->key, [$message])->once()->andReturn(null);
+        $session->shouldReceive('get')->with($this->key)->once()->andReturn([$message]);
 
         // Process
         $stream = new SessionAlertStream($this->key, $translator, $session);
-        $stream->addMessageTranslated('success', $key, $placeholder);
+        $stream->addMessageTranslated('success', $key, $placeholder); // @phpstan-ignore-line
+
+        $translator->shouldReceive('translate')->with($key, $placeholder)->andReturn($result);
+        $this->assertSame($result, $stream->messages()[0]['message']);
     }
 
     /**
@@ -165,11 +143,13 @@ class SessionAlertStreamTest extends TestCase
 
         // Set expectations
         $message = [
-            'type'    => 'success',
-            'message' => 'foo',
+            'type'         => 'success',
+            'message'      => 'foo',
+            'placeholders' => [],
         ];
         $session->shouldReceive('get')->with($this->key)->once()->andReturn([$message]);
         $session->shouldReceive('set')->with($this->key, [])->once()->andReturn(true);
+        $translator->shouldReceive('translate')->with('foo', [])->andReturn('foo');
 
         // Process
         $stream = new SessionAlertStream($this->key, $translator, $session);
@@ -194,20 +174,58 @@ class SessionAlertStreamTest extends TestCase
         $validator->shouldReceive('errors')->once()->andReturn($data);
 
         $message1 = [
-            'type'    => 'danger',
-            'message' => 'Name is required',
+            'type'         => 'danger',
+            'message'      => 'Name is required',
+            'placeholders' => [],
         ];
         $message2 = [
-            'type'    => 'danger',
-            'message' => 'Email should be a valid email address',
+            'type'         => 'danger',
+            'message'      => 'Email should be a valid email address',
+            'placeholders' => [],
         ];
         $session->shouldReceive('get')->with($this->key)->andReturn(false, [$message1], [$message1, $message2]); // Save 1, Save 2, Display both
         $session->shouldReceive('set')->with($this->key, [$message1])->andReturn(null); // Save 1
         $session->shouldReceive('set')->with($this->key, [$message1, $message2])->andReturn(null); // Save 2
+        $translator->shouldReceive('translate')->with($message1['message'], [])->andReturn($message1['message']);
+        $translator->shouldReceive('translate')->with($message2['message'], [])->andReturn($message2['message']);
 
         // Process
         $stream = new SessionAlertStream($this->key, $translator, $session);
         $stream->addValidationErrors($validator);
         $this->assertSame([$message1, $message2], $stream->messages());
+    }
+
+    public function testRealSessionService(): void
+    {
+        // Data
+        $key = 'FOO';
+        $placeholder = ['key' => 'value'];
+        $result = 'Bar';
+        $expectedResult = [
+            'type'         => 'success',
+            'message'      => $result,
+            'placeholders' => $placeholder,
+        ];
+
+        // Setup dependencies
+        $handler = new ArraySessionHandler(60);
+        $session = new Session($handler);
+        $session->destroy();
+        $session->start();
+
+        // Setup Translator
+        $translator = m::mock(Translator::class);
+        $translator->shouldReceive('translate')->with($key, $placeholder)->andReturn($result);
+
+        $stream = new SessionAlertStream($this->key, $translator, $session);
+        $this->assertEmpty($stream->messages());
+        $stream->addMessage('success', $key, $placeholder);
+        $this->assertCount(1, $stream->messages());
+        $this->assertSame([$expectedResult], $stream->messages());
+        $this->assertCount(1, $stream->messages());
+        $stream->addMessage('success', $key, $placeholder);
+        $this->assertCount(2, $stream->messages());
+        $this->assertSame([$expectedResult, $expectedResult], $stream->getAndClearMessages());
+        $this->assertCount(0, $stream->messages());
     }
 }